requirement.txt
hydra-core==1.3.2
scikit-learn==1.7.2
pytorch-crf==0.7.2
torch==2.9.1
torchinfo==1.8.0
tqdm==4.67.1
